♻️ refactor: organize query projection storage
