+++ src/libs/queryProjectionStorage/index.ts
@@ -1,0 +1,2 @@
+export * from './localStorage';
+export type * from './types';
