+++ src/libs/queryProjectionStorage/types.ts
@@ -1,0 +1,15 @@
+export interface QueryProjectionKey {
+  queryKey: string;
+  scope: string;
+}
+
+export interface QueryProjection<T> {
+  data: T;
+  updatedAt: number;
+}
+
+export interface QueryProjectionStorage<T> {
+  get: (key: QueryProjectionKey) => Promise<QueryProjection<T> | undefined>;
+  remove: (key: QueryProjectionKey) => Promise<void>;
+  set: (key: QueryProjectionKey, projection: QueryProjection<T>) => Promise<void>;
+}
